silx.gui.data.NXdataWidgets: move ArrayVolumePlot to its own file
